pyproject.toml
[tool.poetry]
name = "vodafone-r207-sms-mqtt"
version = "0.0.1"
description = "Take incoming SMS from Vodafone R207 MiFi and publish them onto MQTT"
authors = ["Aaron Carson <aaron@aaroncarson.co.uk>"]
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
defusedxml = "^0.7.1"
dicttoxml = "^1.7.16"
xmltodict = "^0.13.0"
paho-mqtt = "^1.6.1"
mariadb = "^1.1.6"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
